Fix Bug #28
When the container is started for the first time we save a copy of the original `/etc/apache2/envvars` file to `/etc/apache2/envvars.orig`. On a container restart we copy the original file `/etc/apache2/envvars.orig` to `/etc/apache2/envvars` and modify it as usual without appending settings to `/etc/apache2/envvars` over and over.

Before starting Apache in the foreground  we ensure that it is not running by explicitly stopping the service.