List repos in v2 registry with paging (#47 and #7)
We can now navigate forward through all images inside a v2 registry. The
default page size is 10 but it can be changed to some other numbers as
well.

The unused repository selector files have been removed.

The repository list directive was moved to the repository list html
template for simplicity.

The repository list no longer says that it searches for repos but
filters repos on the current page.

The bottom delete repos button has been removed in favor of more space
for the pagination controls.

The pagination works as described here:
https://github.com/docker/distribution/blob/master/docs/spec/api.md#pagination
We take the "Link:" header parameter from the response and extract
portions of it to formulate the next request. As usual, each page is
addressible via the URL.

Note that when changing the number of repositories per page we have no
other way but to go to page 1 and show at max the number of repositories
selected.

There's no way to find out how many repos are there so we cannot provide
page links. Only forward navigation is possible.