Add readable host var
This variable will be used to display the hostname of the docker registry
on the tag detail page for the docker pull command.

Since as actual host, usually, the docker container 'registry'
is defined, 'registry' would also show up in this command
even though it's not the actual URL to the docker-registry
from which the image can be pulled.