Update to docker-compose v3
This change set updates the docker-compose file in the develop folder to use
the v3 syntax. This allows us to add a user layer network so that the two
containers can communicate without having to go up the network stack. This
allows us to not have to expose the registery's port (5000) to the host.

This reduces the chance of a port conflict