Change landing page to repositories
Currently the landing page is the home page but that requires the user to make
one mouse click to navigate to the repositories page which is where the action
is. This change set just updates the default landing route from Home to
repositories (with a default page of 20). The Home page still exists but is not
the default location the user lands on when going to the site.

This fixes issues #34