:bug:Fix tag links navigating to home
Some of the test links when clicked would take you to the default page (in this
case the repositories view). This only happened on images which didn't have
a user prefix (i.e. nginx vs. my/nginx).

This happened because the links were built expecting the prefix so the URL
was //nginx but that was not routed properly so the user fell into the catch
other 'otherwise' handler which took them to the repositories view.

This change set fixes that by building the links with URLs that include the
prefix if it exists or without the prefix if the image does not have a user
prefix.